# contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
...
info:
  title: Talosmine Control Plane API
  version: 1.0.0
  description: |
    Control Plane của Talosmine — nguồn sự thật về identity, entitlement và quota.

    **Trạng thái: P1 (bootstrap).** Chỉ có health endpoint. Không có capability nghiệp vụ nào.

    Quy ước xuyên phase:
    - Số lượng (quantity/limit/counter/amount) biểu diễn bằng **decimal string**, không phải
      JSON number — PostgreSQL dùng `bigint` và JSON number mất chính xác ngoài 2^53.
    - Thời gian dùng timestamp có timezone; business time dựa DB clock.
    - Mọi lỗi trả về `ErrorEnvelope` với `code` máy đọc được. Client bắt theo `code`,
      không theo `message`.
  license:
    name: UNLICENSED
servers:
  - url: https://talosmine.localhost
    description: Local dev qua Caddy (DEC-T11)
...
paths:
...
  /v1/me/onboarding:
    get:
      tags: [survey]
      operationId: getOwnOnboarding
      summary: Khảo sát onboarding — cần trả lời hay chưa, và nội dung câu hỏi
      description: |
        Cần phiên đăng nhập: khảo sát gắn với một account cụ thể.

        `required = false` khi account này đã hoàn tất HOẶC đã bỏ qua. BFF gọi endpoint này
        ngay sau khi đổi được phiên ở `/auth/callback` để quyết định có chuyển hướng sang
        màn hình khảo sát hay không.

        Chỉ trả câu hỏi `active` và lựa chọn `active`, kèm nhãn theo `locale` yêu cầu. Lựa
        chọn thiếu bản dịch cho ngôn ngữ đó bị BỎ QUA — cùng quy tắc với điều hướng site.
      security:
        - webSession: []
      parameters:
        - name: locale
          in: query
          required: true
          schema:
            type: string
            enum: [vi, en]
      responses:
        '200':
          description: Trạng thái onboarding và bộ câu hỏi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingSurvey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [survey]
      operationId: submitOwnOnboarding
      summary: Nộp hoặc bỏ qua khảo sát
      description: |
        Cần phiên đăng nhập. Một account nộp ĐÚNG MỘT LẦN — gọi lần hai trả 409.

        `status = 'skipped'` cũng tạo bản ghi và KHÔNG kèm câu trả lời. Ghi lại việc bỏ qua
        là có chủ đích: thiếu nó thì hệ thống không phân biệt được "chưa từng hỏi" với "đã
        hỏi và họ từ chối", nên sẽ hỏi lại mãi.

        SERVER KIỂM LẠI TOÀN BỘ, không tin client — màn hình có thể bị bỏ qua bằng cách gọi
        thẳng endpoint này:
          - mọi `optionKey` phải thuộc đúng `questionKey` và đang `active`;
          - câu `single` chỉ nhận một lựa chọn;
          - câu `multi` phải đủ `minSelect`.
      security:
        - webSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitOnboardingRequest'
      responses:
        '204':
          description: Đã ghi nhận.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Account này đã trả lời khảo sát rồi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/me/account:
    patch:
      tags: [account]
      operationId: updateOwnAccount
      summary: Sửa hồ sơ của chính mình
      description: |
        ALLOWLIST ba trường: `displayName`, `locale`, `timezone`. Trường ngoài danh sách bị
        BỎ QUA im lặng — client cũ gửi thừa vẫn chạy, mà field đó vẫn vô tác dụng.

        `status` chỉ admin đổi qua endpoint riêng có audit. `email` và `emailVerified` do
        IdP sở hữu — tự khai ở đây là tự phong "đã xác minh".

        Chuỗi rỗng nghĩa là XÓA giá trị (lưu NULL), không lưu chuỗi rỗng.
      security:
        - webSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: [string, 'null']
                  maxLength: 100
                locale:
                  type: [string, 'null']
                  maxLength: 35
                timezone:
                  type: [string, 'null']
                  maxLength: 64
      responses:
        '200':
          description: Hồ sơ sau khi cập nhật.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnAccount'
        '400':
          description: Giá trị sai kiểu hoặc vượt độ dài cho phép.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Thiếu phiên hoặc phiên không hợp lệ.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Thiếu hoặc sai CSRF token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags: [account]
      operationId: getOwnAccount
      summary: Xem tài khoản của chính mình
      description: |
        Trả hồ sơ của account gắn với phiên hiện tại. accountId được Control Plane resolve
        từ session token phía server — không có tham số path/query/body nào cho phép hỏi
        account của người khác.
      security:
        - webSession: []
      responses:
        '200':
          description: Hồ sơ account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnAccount'
        '401':
          description: Thiếu phiên, phiên sai, hết hạn hoặc đã thu hồi (fail-closed, không phân biệt lý do).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /v1/admin/survey/summary:
    get:
      tags: [admin]
      operationId: adminGetSurveySummary
      summary: Tổng hợp kết quả khảo sát
      description: |
        Yêu cầu **`survey_response:read`** — permission riêng, KHÔNG phải `content:*`. Sửa
        câu hỏi là việc biên tập; đọc kết quả là truy cập dữ liệu người dùng.

        Đây là thứ trả lời đúng mục đích "thu thập thông tin": mỗi lựa chọn được chọn bao
        nhiêu lần. Danh sách thô chỉ hữu ích khi điều tra một trường hợp cụ thể.

        Không trả `accountId` — tổng hợp không cần biết ai trả lời gì.
      security:
        - webSession: []
      parameters:
        - name: locale
          in: query
          required: true
          description: Ngôn ngữ của nhãn hiển thị trong báo cáo.
          schema:
            type: string
            enum: [vi, en]
      responses:
        '200':
          description: Số liệu tổng hợp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/survey/responses:
    get:
      tags: [admin]
      operationId: adminListSurveyResponses
      summary: Danh sách lượt trả lời (chi tiết)
      description: |
        Yêu cầu **`survey_response:read`**.

        Trả `accountId` để đối chiếu với màn hình quản trị account khi điều tra một trường
        hợp cụ thể. **Đây là dữ liệu cá nhân** — thời hạn lưu thuộc DEC-B11, chưa chốt.

        Phân trang bằng cursor theo `createdAt`, mới nhất trước — cùng khuôn với
        `/v1/admin/audit`.
      security:
        - webSession: []
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Danh sách lượt trả lời, mới nhất trước.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                additionalProperties: false
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SurveyResponseRecord'
                  nextCursor:
                    type: [string, 'null']
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  responses:
    BadRequest:
      description: Dữ liệu vào không hợp lệ (ví dụ thiếu `reason`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Chưa xác thực (thiếu/sai/hết hạn/đã thu hồi phiên).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: |
        Đã xác thực nhưng KHÔNG đủ quyền quản trị. Cố ý không nêu thiếu permission nào —
        tránh vẽ bản đồ phân quyền cho kẻ dò.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  securitySchemes:
    webSession:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Web session token do BFF forward (từ cookie HttpOnly). Control Plane tra hash
        trong `web_sessions`, kiểm chưa hết hạn/chưa thu hồi, rồi resolve accountId.
        Token thô KHÔNG được log. Đây là kênh của NGƯỜI DÙNG — khác `serviceIdentity` (M2M).
    userSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token user do Auth0 phát, được BFF chuyển tiếp.
        **CHƯA dùng ở P1** — wiring Auth0 thật là P2 và phụ thuộc DEC-B03 (tenant thật).
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        M2M token của backend ứng dụng (Data Plane). Mỗi backend có identity RIÊNG,
        audience cụ thể và exact resource scope. Không dùng chung credential.
        **CHƯA dùng ở P1.**
  schemas:
...
    SurveyQuestionKey:
      type: string
      enum: [categories, primary_use, discover_first]
      description: |
        Khoá máy ỔN ĐỊNH. Danh mục ĐÓNG — cấu trúc khảo sát cố định vì code phải có chỗ
        render tương ứng; chỉ nội dung và danh sách lựa chọn sửa được trong `/admin`.
    SurveyIconKey:
      type: string
      description: |
        Khoá icon thuộc danh mục ĐÓNG; code render SVG tương ứng. Không phải URL và không
        phải SVG — CSP theo nonce (DEC-T20) sẽ chặn markup tự nhập, và nới CSP để chiều một
        ô nhập icon 20px là đánh đổi tệ nhất trong kiến trúc này.

        Danh sách này phải khớp CHECK ở migration `0012_survey` VÀ `survey-icons.tsx`.
      enum:
        [
          writing,
          design,
          code,
          video,
          image,
          automation,
          research,
          business,
          chart,
          chat,
          cloud,
          shield,
          sparkle,
          rocket,
          book,
        ]
    OnboardingSurvey:
      type: object
      required: [required, questions]
      additionalProperties: false
      properties:
        required:
          type: boolean
          description: '`false` khi account đã hoàn tất HOẶC đã bỏ qua.'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
    SurveyQuestion:
      type: object
      required: [key, kind, minSelect, title, description, options]
      additionalProperties: false
      properties:
        key:
          $ref: '#/components/schemas/SurveyQuestionKey'
        kind:
          type: string
          enum: [single, multi]
          description: '`multi` render checkbox, `single` render radio.'
        minSelect:
          type: integer
          minimum: 1
          description: Số lựa chọn tối thiểu. Server kiểm lại, không tin client.
        title: { type: string }
        description: { type: [string, 'null'] }
        options:
          type: array
          items:
            $ref: '#/components/schemas/SurveyOption'
    SurveyOption:
      type: object
      required: [key, label, description, icon]
      additionalProperties: false
      properties:
        key:
          type: string
          description: Ổn định trong phạm vi câu hỏi. Đây là thứ client gửi lại khi nộp.
        label: { type: string }
        description:
          type: [string, 'null']
          description: Câu 1 không có; câu 2 và 3 có một dòng mô tả trong ô.
        icon:
          oneOf:
            - $ref: '#/components/schemas/SurveyIconKey'
            - type: 'null'
    SubmitOnboardingRequest:
      type: object
      required: [status, locale]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [completed, skipped]
        locale:
          type: string
          enum: [vi, en]
          description: Ngôn ngữ lúc trả lời — lưu lại vì nhãn lựa chọn có thể đổi về sau.
        answers:
          type: array
          description: Bắt buộc khi `status = completed`; bị bỏ qua khi `skipped`.
          items:
            type: object
            required: [questionKey, optionKeys]
            additionalProperties: false
            properties:
              questionKey:
                $ref: '#/components/schemas/SurveyQuestionKey'
              optionKeys:
                type: array
                minItems: 1
                items: { type: string }

    # ── Survey: quản trị ───────────────────────────────────────────────────────
...
    SurveySummary:
      type: object
      required: [totalResponses, completedCount, skippedCount, questions]
      additionalProperties: false
      properties:
        totalResponses:
          type: integer
          description: Tổng số lượt, gồm cả bỏ qua.
        completedCount: { type: integer }
        skippedCount:
          type: integer
          description: |
            Có giá trị riêng: tỉ lệ bỏ qua cao nghĩa là khảo sát đang cản đường vào sản phẩm.
        questions:
          type: array
          items:
            type: object
            required: [key, title, respondentCount, options]
            additionalProperties: false
            properties:
              key:
                $ref: '#/components/schemas/SurveyQuestionKey'
              title: { type: [string, 'null'] }
              respondentCount:
                type: integer
                description: |
                  Số NGƯỜI đã trả lời câu này — khác tổng `count` của các lựa chọn ở câu
                  `multi`, vì một người chọn nhiều ô. Thiếu số này thì phần trăm tính sai.
              options:
                type: array
                items:
                  type: object
                  required: [key, label, icon, status, count]
                  additionalProperties: false
                  properties:
                    key: { type: string }
                    label: { type: [string, 'null'] }
                    icon:
                      oneOf:
                        - $ref: '#/components/schemas/SurveyIconKey'
                        - type: 'null'
                    status:
                      type: string
                      enum: [draft, active, inactive]
                      description: |
                        Trả cả trạng thái vì một lựa chọn đã `inactive` vẫn còn số đếm lịch
                        sử. Không hiển thị nó thì con số biến mất; không ghi trạng thái thì
                        người đọc tưởng báo cáo hỏng.
                    count: { type: integer }
    SurveyResponseRecord:
      type: object
      required: [id, accountId, status, locale, createdAt, answers]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        accountId:
          type: string
          format: uuid
          description: Để đối chiếu với màn hình quản trị account khi điều tra một trường hợp.
        status: { type: string, enum: [completed, skipped] }
        locale: { type: string, enum: [vi, en] }
        createdAt: { type: string, format: date-time }
        answers:
          type: array
          description: Rỗng khi `status = skipped`.
          items:
            type: object
            required: [questionKey, optionKeys]
            additionalProperties: false
            properties:
              questionKey:
                $ref: '#/components/schemas/SurveyQuestionKey'
              optionKeys:
                type: array
                description: |
                  KHOÁ chứ không phải nhãn — danh sách thô phải đọc được như nhau bất kể bản
                  dịch hiện tại là gì.
                items: { type: string }

    # ── Site content ───────────────────────────────────────────────────────────
...
    OwnAccount:
      type: object
      description: |
        Hồ sơ account trả cho chính user. CỐ Ý không có trường quản trị nội bộ
        (disabled_at, updated_at) — user chỉ thấy hồ sơ, không thấy dữ liệu vận hành.
      required: [id, status, emailVerified, createdAt]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, active, disabled]
        displayName:
          type: [string, 'null']
        email:
          type: [string, 'null']
        emailVerified:
          type: boolean
        locale:
          type: [string, 'null']
        timezone:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      description: |
        Hình dạng lỗi ổn định cho MỌI lỗi rời khỏi API, kể cả lỗi không lường trước.

        Ràng buộc bảo mật: không bao giờ chứa stack trace, secret, connection string
        hay chi tiết nội bộ. Với lỗi 5xx, `message` luôn là thông điệp chung — chi tiết
        thật chỉ đi vào log phía server, truy được bằng `correlationId`.
      required: [code, message, correlationId]
      additionalProperties: false
      properties:
        code:
          type: string
          description: |
            Mã máy đọc được, ổn định, không phụ thuộc câu chữ giao diện.
            Client bắt lỗi theo trường này.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - INTERNAL_ERROR
            - VALIDATION_FAILED
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Thông điệp an toàn để hiển thị.
          maxLength: 500
        correlationId:
          type: string
          format: uuid
          description: |
            Truy vết xuyên BFF -> API -> worker -> log.
            Nhận từ header `X-Correlation-Id` nếu hợp lệ, ngược lại server sinh mới (UUIDv7).
            **Không phải** idempotency key — đừng dùng lẫn.
        details:
          type: object
          description: Chi tiết bổ sung theo allowlist. Không bao giờ chứa dữ liệu nhạy cảm.
          additionalProperties: true
...
security: []